rename all components to component word
