+++ src/components/categories/component.tsx
@@ -1,4 +1,4 @@
-import { Category } from "src/components/category/Category";
+import { Category } from "src/components/category/component";
 import classes from "./styles.module.css";
 import productPageStyles from "src/pages/product-page/styles.module.css";
